Fix: error handling, race condition and business rules
- Map business errors to proper HTTP status codes (error_handlers)
- Fix wrong exception types and validate non-negative account balance
- Lock account rows (with_for_update) and use Decimal to prevent lost updates
- Include incoming transfers in transaction history
- Use TransactionController in transaction routes
- Remove unused auth/security stubs
- Add .env.example and align Dockerfile Python version
- Update README

+++ src/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from database import Base
 
 class Account(Base):
@@ -6,4 +6,4 @@ class Account(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-    balance = Column(Float, default=0.0)
+    balance = Column(Numeric(10, 2), default=0.0)
